+++ Assets/Editor/Tests/Pack1UnarchiverTest.cs
@@ -31,7 +31,8 @@ public class Pack1UnarchiverTest
         string metaString = File.ReadAllText(metaPath);
         Pack1Meta meta = Pack1Meta.Parse(metaString);
 
-        var pack1Unarchiver = new Pack1Unarchiver(archivePath, meta, _tempDir, Key);
+        MapHashExtractedFiles mapHashExtractedFiles = new MapHashExtractedFiles();
+        var pack1Unarchiver = new Pack1Unarchiver(archivePath, meta, _tempDir, mapHashExtractedFiles, Key);
         pack1Unarchiver.Unarchive(new CancellationToken());
 
         Assert.True(Directory.Exists(Path.Combine(_tempDir, "dir")));