+++ Assets/PatchKit Patcher/Scripts/AppUpdater/Commands/RepairFilesCommand.cs
@@ -81,8 +81,9 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
                         DirectoryOperations.CreateDirectory(unarchivePath, cancellationToken);
                     }
 
-                    var downloader = new ChunkedHttpDownloader(packagePath, _resource.ResourceUrls,
-                        _resource.ChunksData, _resource.Size);
+                    MapHashExtractedFiles mapHashExtracted = new MapHashExtractedFiles();
+                    
+                    var downloader = new ChunkedHttpDownloader(packagePath, _resource.ResourceUrls, _resource.ChunksData, _resource.Size);
 
                     long start = entry.Offset.GetValueOrDefault();
                     long end = (start + entry.Size.GetValueOrDefault()) - 1; // Offset by 1 to denote a byte index
@@ -121,8 +122,9 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
                     repairStatus.Progress.Value = 0.0;
 
                     _logger.LogDebug("Unarchiving the package.");
-                    var unarchiver = new Pack1Unarchiver(packagePath, _meta, unarchivePath, _packagePassword,
-                        _unpackingSuffix, effectiveRange);
+
+                    var unarchiver = new Pack1Unarchiver(packagePath, _meta, unarchivePath, mapHashExtracted, _packagePassword, _unpackingSuffix, effectiveRange);
+
                     // allow repair to continue on errors, because after the repair process, all the files must be validated again
                     unarchiver.ContinueOnError = true;
 
@@ -132,9 +134,17 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
                     };
 
                     unarchiver.UnarchiveSingleFile(entry, cancellationToken);
-
-                    EmplaceFile(Path.Combine(unarchivePath, entry.Name + _unpackingSuffix),
-                        Path.Combine(_localData.Path, entry.Name), cancellationToken);
+                    
+                    string nameHash;
+                    if (mapHashExtracted.TryGetHash(entry.Name, out nameHash))
+                    {
+                        EmplaceFile(Path.Combine(unarchivePath, nameHash + _unpackingSuffix),
+                            Path.Combine(_localData.Path, entry.Name), cancellationToken);
+                    }
+                    else
+                    {
+                        throw new Exception(string.Format("Cannot find hash for file {0} in mapHash.", entry.Name));
+                    }
 
                     repairStatus.IsActive.Value = false;
                 });
@@ -184,7 +194,12 @@ namespace PatchKit.Unity.Patcher.AppUpdater.Commands
             {
                 FileOperations.Delete(target, cancellationToken);
             }
-
+#if UNITY_STANDALONE_WIN
+            if (target.Length > 259)
+            {
+                throw new FilePathTooLongException(string.Format("Cannot install file {0}, the destination path length has exceeded Windows path length limit (260).", target)); 
+            }
+#endif
             FileOperations.Move(source, target, cancellationToken);
         }
     }