+++ Assets/PatchKit Patcher/Scripts/Patcher.cs
@@ -14,9 +14,11 @@ using UniRx;
 using UnityEngine;
 using CancellationToken = PatchKit.Unity.Patcher.Cancellation.CancellationToken;
 using System.IO;
+using PatchKit.Api.Models.Main;
 using PatchKit.Network;
 using PatchKit.Unity.Patcher.AppData;
 using PatchKit.Unity.Patcher.AppData.FileSystem;
+using PatchKit.Unity.Patcher.AppData.Local;
 using PatchKit.Unity.Patcher.AppUpdater.Status;
 using PatchKit.Unity.Patcher.UI;
 using PatchKit.Unity.Patcher.UI.NewUI;
@@ -40,7 +42,9 @@ namespace PatchKit.Unity.Patcher
             InstallApp,
             InstallAppAutomatically,
             CheckForAppUpdates,
-            CheckForAppUpdatesAutomatically
+            CheckForAppUpdatesAutomatically,
+            UninstallApp,
+            VerifyFiles
         }
 
         private static readonly DebugLogger DebugLogger = new DebugLogger(typeof(Patcher));
@@ -284,6 +288,8 @@ namespace PatchKit.Unity.Patcher
             {
                 StartThread();
             }
+
+            gameObject.AddComponent<DebugMenu>();
         }
 
         /// <summary>
@@ -617,6 +623,11 @@ namespace PatchKit.Unity.Patcher
                     _lockFileStream = File.Open(lockFilePath, FileMode.Append);
                     DebugLogger.Log("Lock file open success");
                 }
+                catch (UnauthorizedAccessException exception)
+                {
+                    DebugLogger.LogError("Patcher does not have permission to create the .lock file");
+                    DebugLogger.LogException(exception);
+                }
                 catch
                 {
                     throw new MultipleInstancesException("Another instance of Patcher spotted");
@@ -785,6 +796,12 @@ namespace PatchKit.Unity.Patcher
                     case UserDecision.CheckForAppUpdates:
                         ThreadUpdateApp(false, cancellationToken);
                         break;
+                    case UserDecision.UninstallApp:
+                        ThreadUninstallApp(cancellationToken);
+                        break;
+                    case UserDecision.VerifyFiles:
+                        ThreadVerifyAllAppFiles(cancellationToken);
+                        break;
                 }
 
                 DebugLogger.Log(string.Format("User decision {0} execution done.", _userDecision));
@@ -830,6 +847,11 @@ namespace PatchKit.Unity.Patcher
             {
                 DebugLogger.LogException(e);
                 ThreadDisplayError(PatcherError.CannotRepairDiskFilesException(), cancellationToken);
+            }
+            catch (FilePathTooLongException e)
+            {
+                DebugLogger.LogException(e);
+                ThreadDisplayError(PatcherError.FilePathTooLong(), cancellationToken);
             }
             catch (ThreadInterruptedException)
             {
@@ -921,6 +943,7 @@ namespace PatchKit.Unity.Patcher
             {
                 _appInfo.Value = _app.RemoteMetaData.GetAppInfo(!automatically, _updateAppCancellationTokenSource.Token);
                 _remoteVersionId.Value = _app.GetLatestVersionId(!automatically, _updateAppCancellationTokenSource.Token);
+                
                 if (_app.IsFullyInstalled())
                 {
                     _localVersionId.Value = _app.GetInstalledVersionId();
@@ -936,6 +959,86 @@ namespace PatchKit.Unity.Patcher
                     {
                         appUpdater.Update(_updateAppCancellationTokenSource.Token);
                         _wasUpdateSuccessfulOrNotNecessary = true;
+                    }
+
+                    AppVersion latestAppVersion =
+                        _app.RemoteMetaData.GetAppVersionInfo(_remoteVersionId.Value.Value, false, cancellationToken);
+                    _app.LocalMetaData.SetMainExecutableAndArgs(latestAppVersion.MainExecutable, latestAppVersion.MainExecutableArgs);
+                }
+                catch (OperationCanceledException)
+                {
+                    PatcherStatistics.DispatchSendEvent(PatcherStatistics.Event.PatcherCanceled);
+
+                    throw;
+                }
+                finally
+                {
+                    _state.Value = PatcherState.None;
+
+                    _updaterStatus.Value = null;
+                    _updateAppCancellationTokenSource = null;
+                }
+            }
+        }
+        
+        private void ThreadVerifyAllAppFiles(CancellationToken cancellationToken)
+        {
+            // TODO: Introduce here a new state
+            _state.Value = PatcherState.UpdatingApp;
+
+            _updateAppCancellationTokenSource = new PatchKit.Unity.Patcher.Cancellation.CancellationTokenSource();
+   
+            using (cancellationToken.Register(() => _updateAppCancellationTokenSource.Cancel()))
+            {
+                _isAppInstalled.Value = false;
+                
+                var appUpdater = new AppUpdater.AppUpdater(new AppUpdaterContext(_app, _configuration.AppUpdaterConfiguration));
+
+                try
+                {
+                    _updaterStatus.Value = appUpdater.Status;
+
+                    using (_updaterStatus.Take(1).Subscribe((status) => _state.Value = PatcherState.UpdatingApp))
+                    {
+                        appUpdater.VerifyFiles(cancellationToken);
+                    }
+                }
+                catch (OperationCanceledException)
+                {
+                    PatcherStatistics.DispatchSendEvent(PatcherStatistics.Event.PatcherCanceled);
+
+                    throw;
+                }
+                finally
+                {
+                    _state.Value = PatcherState.None;
+
+                    _updaterStatus.Value = null;
+                    _updateAppCancellationTokenSource = null;
+                }
+            }
+        }
+        
+        private void ThreadUninstallApp(CancellationToken cancellationToken)
+        {
+            // TODO: Introduce here a new state
+            _state.Value = PatcherState.UpdatingApp;
+
+            _updateAppCancellationTokenSource = new PatchKit.Unity.Patcher.Cancellation.CancellationTokenSource();
+   
+            using (cancellationToken.Register(() => _updateAppCancellationTokenSource.Cancel()))
+            {
+                _isAppInstalled.Value = false;
+                
+                var appUpdater = new AppUpdater.AppUpdater(new AppUpdaterContext(_app, _configuration.AppUpdaterConfiguration));
+
+                try
+                {
+                    _updaterStatus.Value = appUpdater.Status;
+
+                    using (_updaterStatus.Take(1).Subscribe((status) => _state.Value = PatcherState.UpdatingApp))
+                    {
+                        appUpdater.Uninstall(cancellationToken);
                     }
                 }
                 catch (OperationCanceledException)